+++ HW13WebService/src/main/java/ru/otus/core/dao/UserDao.java
@@ -1,0 +1,15 @@
+package ru.otus.core.dao;
+
+import ru.otus.core.model.User;
+import ru.otus.core.sessionmanager.SessionManager;
+
+import java.util.Optional;
+
+public interface UserDao {
+    Optional<User> findById(long id);
+    long saveUser(User user);
+    long saveOrUpdate(User user);
+    SessionManager getSessionManager();
+    Optional<User> findByLogin(String login);
+    Optional<User> findRandomUser();
+}
