+++ HW13WebService/src/main/java/ru/otus/core/dao/UserDaoException.java
@@ -1,0 +1,7 @@
+package ru.otus.core.dao;
+
+public class UserDaoException extends RuntimeException {
+    public UserDaoException(Exception ex) {
+        super(ex);
+    }
+}
