+++ HW13WebService/src/main/java/ru/otus/core/service/DbServiceException.java
@@ -1,0 +1,7 @@
+package ru.otus.core.service;
+
+public class DbServiceException extends RuntimeException {
+    public DbServiceException(Exception e) {
+        super(e);
+    }
+}
