+++ HW13WebService/src/main/java/ru/otus/core/service/DbServiceUser.java
@@ -1,0 +1,11 @@
+package ru.otus.core.service;
+
+import ru.otus.core.model.User;
+
+import java.util.Optional;
+
+public interface DbServiceUser {
+    long saveUser(User user);
+    Optional<User> getUser(long id);
+    long saveOrUpdate(User user);
+}
