+++ HW13WebService/src/main/java/ru/otus/core/sessionmanager/DatabaseSession.java
@@ -1,0 +1,4 @@
+package ru.otus.core.sessionmanager;
+
+public interface DatabaseSession {
+}
