+++ HW13WebService/src/main/java/ru/otus/core/sessionmanager/SessionManager.java
@@ -1,0 +1,10 @@
+package ru.otus.core.sessionmanager;
+
+public interface SessionManager extends AutoCloseable{
+    void beginSession();
+    void commitSession();
+    void rollbackSession();
+    void close();
+
+    DatabaseSession getCurrentSession();
+}
