+++ HW13WebService/src/main/java/ru/otus/core/sessionmanager/SessionManagerException.java
@@ -1,0 +1,11 @@
+package ru.otus.core.sessionmanager;
+
+public class SessionManagerException extends RuntimeException {
+    public SessionManagerException(String msg) {
+        super(msg);
+    }
+
+    public SessionManagerException(Exception ex) {
+        super(ex);
+    }
+}
