+++ HW13WebService/src/main/java/ru/otus/hibernate/dao/UserDaoHibernate.java
@@ -1,0 +1,87 @@
+package ru.otus.hibernate.dao;
+
+import org.hibernate.Session;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ru.otus.core.dao.UserDao;
+import ru.otus.core.dao.UserDaoException;
+import ru.otus.core.model.User;
+import ru.otus.core.sessionmanager.SessionManager;
+import ru.otus.hibernate.sessionmanager.DatabaseSessionHibernate;
+import ru.otus.hibernate.sessionmanager.SessionManagerHibernate;
+
+import java.util.Optional;
+import java.util.Random;
+
+
+public class UserDaoHibernate implements UserDao {
+    private static Logger logger = LoggerFactory.getLogger(UserDaoHibernate.class);
+    private final SessionManagerHibernate sessionManager;
+
+    public UserDaoHibernate(SessionManagerHibernate sessionManager) {
+        this.sessionManager = sessionManager;
+    }
+
+    @Override
+    public Optional<User> findById(long id) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            return Optional.ofNullable(currentSession.getHibernateSession().find(User.class, id));
+        } catch (Exception ex) {
+            logger.error(ex.getMessage(), ex);
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public long saveUser(User user) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            Session hibernateSession = currentSession.getHibernateSession();
+            if (user.getId() > 0) {
+                hibernateSession.merge(user);
+            } else {
+                hibernateSession.persist(user);
+            }
+            return user.getId();
+        } catch (Exception ex) {
+            logger.error(ex.getMessage(), ex);
+            throw new UserDaoException(ex);
+        }
+    }
+
+    @Override
+    public long saveOrUpdate(User user) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            Session hibernateSession = currentSession.getHibernateSession();
+            hibernateSession.saveOrUpdate(user);
+            return user.getId();
+        } catch (Exception ex) {
+            logger.error(ex.getMessage(), ex);
+            throw new UserDaoException(ex);
+        }
+    }
+
+    @Override
+    public Optional<User> findByLogin(String login) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            return Optional.ofNullable(currentSession.getHibernateSession().find(User.class, login));
+        } catch (Exception ex) {
+            logger.error(ex.getMessage(), ex);
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public Optional<User> findRandomUser() {
+        Random r = new Random();
+        //return user.values().stream().skip(r.nextInt(users.size() - 1)).findFirst();
+    }
+
+    @Override
+    public SessionManager getSessionManager() {
+        return sessionManager;
+    }
+}
