+++ HW13WebService/src/main/java/ru/otus/hibernate/sessionmanager/DatabaseSessionHibernate.java
@@ -1,0 +1,30 @@
+package ru.otus.hibernate.sessionmanager;
+
+import org.hibernate.Session;
+import org.hibernate.Transaction;
+import ru.otus.core.sessionmanager.DatabaseSession;
+
+public class DatabaseSessionHibernate  implements DatabaseSession {
+    private final Session session;
+    private final Transaction transaction;
+
+    public DatabaseSessionHibernate(Session session) {
+        this.session = session;
+        this.transaction = session.beginTransaction();
+    }
+
+    public Session getHibernateSession() {
+        return session;
+    }
+
+    public Transaction getTransaction() {
+        return transaction;
+    }
+
+    public void close() {
+        if(transaction.isActive()){
+            transaction.commit();
+        }
+        session.close();
+    }
+}
