--- HW13WebService/src/main/java/ru/otus/hw10/core/dao/UserDaoException.java
@@ -1,7 +1,0 @@
-package ru.otus.hw10.core.dao;
-
-public class UserDaoException extends RuntimeException {
-    public UserDaoException(Exception ex) {
-        super(ex);
-    }
-}
