--- HW13WebService/src/main/java/ru/otus/hw10/core/model/User.java
@@ -1,96 +1,0 @@
-package ru.otus.hw10.core.model;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-@Entity
-@Table(name = "app_user")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private long id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "age")
-    private int age;
-
-    @OneToOne(targetEntity = Address.class, cascade = CascadeType.ALL)
-    @JoinColumn(name = "r_address_id")
-    private Address address;
-
-    @OneToMany(targetEntity = Phone.class, cascade = CascadeType.ALL, fetch = FetchType.EAGER, mappedBy = "app_user")
-    private List<Phone> phones;
-
-    @Column(name = "login")
-    private String login;
-
-    @Column(name = "password")
-    private String password;
-
-    public User() {
-    }
-
-    public User(long id, String name, String login, String password) {
-        this.id = id;
-        this.name = name;
-        this.login = login;
-        this.password = password;
-        this.address = null;
-        this.phones = new ArrayList<>();
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public Address getUserAddress() {
-        return address;
-    }
-
-    public void setUserAddress(Address userAddress) {
-        this.address = userAddress;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public List<Phone> getPhones() {
-        return phones;
-    }
-
-    public void addPhone(Phone phone) {
-        phones.add(phone);
-        phone.setUser(this);
-    }
-}
