--- HW13WebService/src/main/java/ru/otus/hw10/core/service/DbServiceException.java
@@ -1,7 +1,0 @@
-package ru.otus.hw10.core.service;
-
-public class DbServiceException extends RuntimeException {
-    public DbServiceException(Exception e) {
-        super(e);
-    }
-}
