--- HW13WebService/src/main/java/ru/otus/hw10/core/sessionmanager/DatabaseSession.java
@@ -1,4 +1,0 @@
-package ru.otus.hw10.core.sessionmanager;
-
-public interface DatabaseSession {
-}
