--- HW13WebService/src/main/java/ru/otus/hw10/core/sessionmanager/SessionManager.java
@@ -1,10 +1,0 @@
-package ru.otus.hw10.core.sessionmanager;
-
-public interface SessionManager extends AutoCloseable{
-    void beginSession();
-    void commitSession();
-    void rollbackSession();
-    void close();
-
-    DatabaseSession getCurrentSession();
-}
