--- HW13WebService/src/main/java/ru/otus/hw10/core/sessionmanager/SessionManagerException.java
@@ -1,11 +1,0 @@
-package ru.otus.hw10.core.sessionmanager;
-
-public class SessionManagerException extends RuntimeException {
-    public SessionManagerException(String msg) {
-        super(msg);
-    }
-
-    public SessionManagerException(Exception ex) {
-        super(ex);
-    }
-}
