--- HW13WebService/src/main/java/ru/otus/hw10/hibernate/sessionmanager/SessionManagerHibernate.java
@@ -1,97 +1,0 @@
-package ru.otus.hw10.hibernate.sessionmanager;
-
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.Transaction;
-import ru.otus.hw10.core.sessionmanager.DatabaseSession;
-import ru.otus.hw10.core.sessionmanager.SessionManager;
-import ru.otus.hw10.core.sessionmanager.SessionManagerException;
-
-public class SessionManagerHibernate implements SessionManager {
-
-    private DatabaseSessionHibernate databaseSession;
-    private final SessionFactory sessionFactory;
-
-    public SessionManagerHibernate(SessionFactory sessionFactory) {
-        if (sessionFactory == null) {
-            throw new SessionManagerException("SessionFactory is null");
-        }
-        this.sessionFactory = sessionFactory;
-    }
-
-    @Override
-    public void beginSession() {
-        try {
-            databaseSession = new DatabaseSessionHibernate(sessionFactory.openSession());
-        } catch (Exception ex) {
-            throw new SessionManagerException(ex);
-        }
-    }
-
-    @Override
-    public void commitSession() {
-        checkSessionAndTransaction();
-        try {
-            databaseSession.getTransaction().commit();
-            databaseSession.getHibernateSession().close();
-        } catch (Exception ex) {
-            throw new SessionManagerException(ex);
-        }
-    }
-
-    @Override
-    public void rollbackSession() {
-        checkSessionAndTransaction();
-        try {
-            databaseSession.getTransaction().rollback();
-            databaseSession.getHibernateSession().close();
-        } catch (Exception ex) {
-            throw new SessionManagerException(ex);
-        }
-    }
-
-    @Override
-    public void close() {
-        if (databaseSession == null) {
-            return;
-        }
-        Session session = databaseSession.getHibernateSession();
-        if (session == null || !session.isConnected()) {
-            return;
-        }
-
-        Transaction transaction = databaseSession.getTransaction();
-        if (transaction == null || !transaction.isActive()) {
-            return;
-        }
-
-        try {
-            databaseSession.close();
-            databaseSession = null;
-        } catch (Exception ex) {
-            throw new SessionManagerException(ex);
-        }
-    }
-
-    @Override
-    public DatabaseSessionHibernate getCurrentSession() {
-        checkSessionAndTransaction();
-        return databaseSession;
-    }
-
-    private void checkSessionAndTransaction() {
-        if (databaseSession == null) {
-            throw new SessionManagerException("DatabaseSession not opened ");
-        }
-
-        Session session = databaseSession.getHibernateSession();
-        if (session == null || !session.isConnected()) {
-            throw new SessionManagerException("Session not opened ");
-        }
-
-        Transaction transaction = databaseSession.getTransaction();
-        if (transaction == null || !transaction.isActive()) {
-            throw new SessionManagerException("Transaction not opened ");
-        }
-    }
-}
