--- HW13WebService/src/main/java/ru/otus/hw12/WebServerDemo.java
@@ -1,7 +1,0 @@
-package ru.otus.hw12;
-
-public class WebServerDemo {
-    public static void main(String[] args) {
-
-    }
-}
