--- HW13WebService/src/main/java/ru/otus/hw12/server/UsersWebServer.java
@@ -1,7 +1,0 @@
-package ru.otus.hw12.server;
-
-public interface UsersWebServer {
-    void start() throws Exception;
-    void join() throws Exception;
-    void stop() throws Exception;
-}
