--- HW13WebService/src/main/java/ru/otus/hw12/services/UserAuthService.java
@@ -1,5 +1,0 @@
-package ru.otus.hw12.services;
-
-public interface UserAuthService {
-    boolean authenticate(String login, String password);
-}
