--- HW13WebService/src/main/java/ru/otus/hw12/services/UserAuthServiceImpl.java
@@ -1,16 +1,0 @@
-package ru.otus.hw12.services;
-
-import ru.otus.hw10.core.dao.UserDao;
-
-public class UserAuthServiceImpl implements  UserAuthService {
-    private final UserDao userDao;
-
-    public UserAuthServiceImpl(UserDao userDao) {
-        this.userDao = userDao;
-    }
-
-    @Override
-    public boolean authenticate(String login, String password) {
-        return userDao.findByLogin(login).map(user -> user.getPassword().equals(password)).orElse(false);
-    }
-}
