--- HW13WebService/src/main/java/ru/otus/hw12/servlet/UserApiServlet.java
@@ -1,38 +1,0 @@
-package ru.otus.hw12.servlet;
-
-import com.google.gson.Gson;
-import ru.otus.hw10.core.dao.UserDao;
-import ru.otus.hw10.core.model.User;
-
-import javax.servlet.ServletOutputStream;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class UserApiServlet extends HttpServlet {
-    private static final int ID_PATH_PARAM_POSITION = 1;
-
-    private final UserDao userDao;
-    private final Gson gson;
-
-    public UserApiServlet(UserDao userDao, Gson gson) {
-        this.userDao = userDao;
-        this.gson = gson;
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        User user = userDao.findById(extractIdFromRequest(request)).orElse(null);
-
-        response.setContentType("application/json;charset=UTF-8");
-        ServletOutputStream out = response.getOutputStream();
-        out.print(gson.toJson(user));
-    }
-
-    private long  extractIdFromRequest(HttpServletRequest request) {
-        String[] path = request.getPathInfo().split("/");
-        String id = (path.length>1)? path[ID_PATH_PARAM_POSITION]: String.valueOf(-1);
-        return Long.parseLong(id);
-    }
-}
