+++ HW13WebService/src/main/java/ru/otus/webserviceHW/WebServerDemo.java
@@ -1,0 +1,7 @@
+package ru.otus.webserviceHW;
+
+public class WebServerDemo {
+    public static void main(String[] args) {
+
+    }
+}
