+++ HW13WebService/src/main/java/ru/otus/webserviceHW/server/UsersWebServer.java
@@ -1,0 +1,7 @@
+package ru.otus.webserviceHW.server;
+
+public interface UsersWebServer {
+    void start() throws Exception;
+    void join() throws Exception;
+    void stop() throws Exception;
+}
