+++ HW13WebService/src/main/java/ru/otus/webserviceHW/server/UsersWebServerImpl.java
@@ -1,0 +1,80 @@
+package ru.otus.webserviceHW.server;
+
+import com.google.gson.Gson;
+import org.eclipse.jetty.server.Handler;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.handler.HandlerList;
+import org.eclipse.jetty.server.handler.ResourceHandler;
+import org.eclipse.jetty.servlet.ServletContextHandler;
+import org.eclipse.jetty.servlet.ServletHolder;
+import ru.otus.core.dao.UserDao;
+import ru.otus.webserviceHW.helpers.FileSystemHelper;
+import ru.otus.webserviceHW.services.TemplateProcessor;
+import ru.otus.webserviceHW.servlet.UsersApiServlet;
+import ru.otus.webserviceHW.servlet.UsersServlet;
+
+public class UsersWebServerImpl implements UsersWebServer{
+    private static final String START_PAGE_NAME = "index.html";
+    private static final String COMMON_RESOURCES_DIR = "static";
+
+    private final UserDao userDao;
+    private final Gson gson;
+    protected final TemplateProcessor templateProcessor;
+    private final Server server;
+
+    public UsersWebServerImpl(int port, UserDao userDao, Gson gson, TemplateProcessor templateProcessor ) {
+        this.userDao = userDao;
+        this.gson = gson;
+        this.templateProcessor = templateProcessor;
+        server = new Server(port);
+    }
+
+    @Override
+    public void start() throws Exception {
+        server.start();
+    }
+
+    @Override
+    public void join() throws Exception {
+        server.join();
+    }
+
+    @Override
+    public void stop() throws Exception {
+        server.stop();
+    }
+
+
+    private Server initContext() {
+
+        ResourceHandler resourceHandler = createResourceHandler();
+        ServletContextHandler servletContextHandler = createServletContextHandler();
+
+        HandlerList handlers = new HandlerList();
+        handlers.addHandler(resourceHandler);
+        handlers.addHandler(applySecurity(servletContextHandler, "/users", "/api/user/*"));
+
+
+        server.setHandler(handlers);
+        return server;
+    }
+
+    protected Handler applySecurity(ServletContextHandler servletContextHandler, String ...paths) {
+        return servletContextHandler;
+    }
+
+    private ResourceHandler createResourceHandler() {
+        ResourceHandler resourceHandler = new ResourceHandler();
+        resourceHandler.setDirectoriesListed(false);
+        resourceHandler.setWelcomeFiles(new String[]{START_PAGE_NAME});
+        resourceHandler.setResourceBase(FileSystemHelper.localFileNameOrResourceNameToFullPath(COMMON_RESOURCES_DIR));
+        return resourceHandler;
+    }
+
+    private ServletContextHandler createServletContextHandler(){
+        ServletContextHandler servletContextHandler = new ServletContextHandler(ServletContextHandler.SESSIONS);
+        servletContextHandler.addServlet(new ServletHolder(new UsersServlet(templateProcessor, userDao)), "/users");
+        servletContextHandler.addServlet(new ServletHolder(new UsersApiServlet(userDao, gson)), "/api/user/*");
+        return servletContextHandler;
+    }
+}
