+++ HW13WebService/src/main/java/ru/otus/webserviceHW/services/TemplateProcessor.java
@@ -1,0 +1,8 @@
+package ru.otus.webserviceHW.services;
+
+import java.io.IOException;
+import java.util.Map;
+
+public interface TemplateProcessor {
+    String getPage(String filename, Map<String, Object> data) throws IOException;
+}
