+++ HW13WebService/src/main/java/ru/otus/webserviceHW/services/UserAuthService.java
@@ -1,0 +1,5 @@
+package ru.otus.webserviceHW.services;
+
+public interface UserAuthService {
+    boolean authenticate(String login, String password);
+}
