+++ HW13WebService/src/main/java/ru/otus/webserviceHW/services/UserAuthServiceImpl.java
@@ -1,0 +1,16 @@
+package ru.otus.webserviceHW.services;
+
+import ru.otus.core.dao.UserDao;
+
+public class UserAuthServiceImpl implements  UserAuthService {
+    private final UserDao userDao;
+
+    public UserAuthServiceImpl(UserDao userDao) {
+        this.userDao = userDao;
+    }
+
+    @Override
+    public boolean authenticate(String login, String password) {
+        return userDao.findByLogin(login).map(user -> user.getPassword().equals(password)).orElse(false);
+    }
+}
