+++ HW13WebService/src/main/java/ru/otus/webserviceHW/servlet/LoginServlet.java
@@ -1,0 +1,49 @@
+package ru.otus.webserviceHW.servlet;
+
+import ru.otus.webserviceHW.services.TemplateProcessor;
+import ru.otus.webserviceHW.services.UserAuthService;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import java.io.IOException;
+import java.util.Collections;
+
+public class LoginServlet extends HttpServlet {
+
+    private static final String PARAM_LOGIN = "login";
+    private static final String PARAM_PASSWORD = "password";
+    private static final int MAX_INACTIVE_INTERVAL = 30;
+    private static final String LOGIN_PAGE_TEMPLATE = "login.html";
+
+    private final TemplateProcessor templateProcessor;
+    private final UserAuthService userAuthService;
+
+    public LoginServlet(TemplateProcessor templateProcessor, UserAuthService userAuthService) {
+        this.userAuthService = userAuthService;
+        this.templateProcessor = templateProcessor;
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html");
+        response.getWriter().println(templateProcessor.getPage(LOGIN_PAGE_TEMPLATE, Collections.emptyMap()));
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        String name = request.getParameter(PARAM_LOGIN);
+        String password = request.getParameter(PARAM_PASSWORD);
+
+        if(userAuthService.authenticate(name, password)){
+            HttpSession session = request.getSession();
+            session.setMaxInactiveInterval(MAX_INACTIVE_INTERVAL);
+            response.sendRedirect("/users");
+        }else {
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        }
+    }
+}
